requirements.txt
anyio==3.6.2
autopep8==2.0.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
fastapi-sessions==0.3.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
orjson==3.8.12
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==1.10.7
pyparsing==3.0.9
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
